uporządkowanie systemu logowania/rejestracji
