+++ main.py
@@ -6,8 +6,9 @@
 from flask import Flask
 
 # importy nasze
-from views import pages, login_manager
+from views import pages
 from views_admin import pages_admin
+from views_login_system import pages_log_sys, login_manager
 from mail_sending import mail
 from db_models import db
 from config import DB, APP, EMAIL
@@ -26,6 +27,7 @@ app.config['MAIL_DEFAULT_SENDER'] = EMAIL.DEFAULT_SENDER
 
 app.register_blueprint(pages)
 app.register_blueprint(pages_admin)
+app.register_blueprint(pages_log_sys)
 login_manager.init_app(app)
 mail.init_app(app)
 db.init_app(app)
