+++ templates/admin_page.html
@@ -45,6 +45,7 @@
         <input class="send_data" type="submit" value="Wyślij">
     </form>
 </div>
+<!-- inna opcja -->
 <div id="" class="admin_option_button"><span>???</span></div>
 <script type="text/javascript" src="{{ url_for('static', filename='scripts/admin.js') }} "></script>
 {% endblock %}
