+++ templates/navbar.html
@@ -10,7 +10,7 @@
     </div>
     {% if current_user.is_authenticated %}
     <div class="nav_button_right">
-        <a class="nav_right" href="{{url_for('pages.logout')}}">Wyloguj</a>
+        <a class="nav_right" href="{{url_for('pages_log_sys.logout')}}">Wyloguj</a>
     </div>
     <div class="nav_button_right" style="min-width: 100px">
         <a class="nav_log_in_as" href="{{ url_for('pages.user_page') }}">
@@ -24,10 +24,10 @@
     </div>
     {% else %}
     <div class="nav_button_right">
-        <a class="nav_right" href="{{url_for('pages.register')}}">Rejestruj</a>
+        <a class="nav_right" href="{{url_for('pages_log_sys.register')}}">Rejestruj</a>
     </div>
     <div class="nav_button_right">
-        <a class="nav_right" href="{{url_for('pages.login')}}">Zaloguj</a>
+        <a class="nav_right" href="{{url_for('pages_log_sys.login')}}">Zaloguj</a>
     </div>
     {% endif %}
     <div style="clear: both"></div>
