+++ templates/user_settings.html
@@ -16,7 +16,7 @@
 {% from "_formhelpers.html" import render_field %}
 <form method=post name="form_box" onsubmit="return validate()">
 <div id="register_box">
-{% if request.url_rule.endpoint == 'pages.register' %}
+{% if request.url_rule.endpoint == 'pages_log_sys.register' %}
 <div class="input_category">Dane wymagane do rejestracji:</div>
 <!-- input do podania emaila -->
 <div>{{ render_field(form_email.email, class="input_field") }}</div>
@@ -36,7 +36,7 @@
 <div class="error_msg" id="old_pw_field_error"></div>
 {% endif %}
 
-{% if request.url_rule.endpoint in ['pages.register', 'pages.user_set_pw']  %}
+{% if request.url_rule.endpoint in ['pages_log_sys.register', 'pages.user_set_pw']  %}
 <!-- rejestracja uzytkownika/zmiana hasła, podanie hasła -->
 <div>{{ render_field(form_pw.password, class="input_field") }}</div>
 <!-- rejestracja uzytkownika/zmiana hasła, powtórz hasło -->
@@ -48,7 +48,7 @@
 <div class="input_category">Dane dodatkowe:</div>
 {% endif %}
 
-{% if request.url_rule.endpoint in ['pages.register', 'pages.user_set_data']  %}
+{% if request.url_rule.endpoint in ['pages_log_sys.register', 'pages.user_set_data']  %}
 <div style="float: left">{{ render_field(form_data.name, class="input_field_1_2") }}</div>
 <div style="float: left">{{ render_field(form_data.last_name, class="input_field_1_2") }}</div>
 <div style="float: left">{{ render_field(form_data.city, class="input_field_3_4") }}</div>
@@ -59,9 +59,9 @@
 <div style="clear: both"></div>
 {% endif %}
 
-{% if request.url_rule.endpoint == 'pages.register' %}
+{% if request.url_rule.endpoint == 'pages_log_sys.register' %}
 <input class ="log_reg_button" type=submit value="Rejestruj">
-<a class="small_redirect_text" href="{{url_for('pages.login')}}">Masz już konto? Zaloguj się!</a>
+<a class="small_redirect_text" href="{{url_for('pages_log_sys.login')}}">Masz już konto? Zaloguj się!</a>
 {% endif %}
 
 {% if request.url_rule.endpoint in ['pages.user_set_pw', 'pages.user_set_data'] %}
